pi/requirements.txt
RPi.GPIO
keyboard
Flask